+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_compose_specs.py
@@ -4,7 +4,7 @@ from typing import Any, Final, TypeAlias, TypedDict
 
 from common_library.json_serialization import json_dumps
 from fastapi.applications import FastAPI
-from models_library.docker import DockerGenericTag, StandardSimcoreDockerLabels
+from models_library.docker import DockerGenericTag
 from models_library.products import ProductName
 from models_library.projects import ProjectID
 from models_library.projects_nodes_io import NodeID
@@ -14,6 +14,7 @@ from models_library.service_settings_labels import (
     SimcoreServiceLabels,
 )
 from models_library.services import ServiceKey, ServiceVersion
+from models_library.services_metadata_runtime import SimcoreContainerLabels
 from models_library.services_resources import (
     DEFAULT_SINGLE_SERVICE_NAME,
     ResourcesDict,
@@ -129,7 +130,7 @@ def _update_paths_mappings(
         env_vars["DY_SIDECAR_PATH_INPUTS"] = f"{path_mappings.inputs_path}"
         env_vars["DY_SIDECAR_PATH_OUTPUTS"] = f"{path_mappings.outputs_path}"
         env_vars["DY_SIDECAR_STATE_PATHS"] = (
-            f"{json_dumps( { f'{p}' for p in path_mappings.state_paths } )}"
+            f"{json_dumps({f'{p}' for p in path_mappings.state_paths})}"
         )
 
         service_content["environment"] = _EnvironmentSection.export_as_list(env_vars)
@@ -202,7 +203,7 @@ def _update_resource_limits_and_reservations(
         ]
 
         resource_limits = [
-            f"{CPU_RESOURCE_LIMIT_KEY}={int(nano_cpu_limits*_NANO)}",
+            f"{CPU_RESOURCE_LIMIT_KEY}={int(nano_cpu_limits * _NANO)}",
             f"{MEM_RESOURCE_LIMIT_KEY}={mem_limits}",
         ]
 
@@ -241,7 +242,7 @@ def _update_container_labels(
             spec_service_key, default_limits
         )
 
-        label_keys = StandardSimcoreDockerLabels.model_validate(
+        label_keys = SimcoreContainerLabels.model_validate(
             {
                 "user_id": user_id,
                 "project_id": project_id,
