+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_service_specs/sidecar.py
@@ -8,14 +8,16 @@ from models_library.aiodocker_api import AioDockerServiceSpec
 from models_library.basic_types import BootModeEnum, PortInt
 from models_library.callbacks_mapping import CallbacksMapping
 from models_library.docker import (
-    DOCKER_TASK_EC2_INSTANCE_TYPE_PLACEMENT_CONSTRAINT_KEY,
     DockerLabelKey,
     DockerPlacementConstraint,
-    StandardSimcoreDockerLabels,
-    to_simcore_runtime_docker_label_key,
 )
 from models_library.resource_tracker import HardwareInfo
 from models_library.service_settings_labels import SimcoreServiceSettingsLabel
+from models_library.services_metadata_runtime import (
+    DOCKER_TASK_EC2_INSTANCE_TYPE_PLACEMENT_CONSTRAINT_KEY,
+    SimcoreContainerLabels,
+    to_simcore_runtime_docker_label_key,
+)
 from pydantic import ByteSize, TypeAdapter
 from servicelib.rabbitmq import RabbitMQRPCClient
 from servicelib.rabbitmq.rpc_interfaces.efs_guardian import efs_manager
@@ -460,18 +462,16 @@ async def get_dynamic_sidecar_spec(  # pylint:disable=too-many-arguments# noqa: 
         scheduler_data.product_name is not None
     ), "ONLY for legacy. This function should not be called with product_name==None"  # nosec
 
-    standard_simcore_docker_labels: dict[DockerLabelKey, str] = (
-        StandardSimcoreDockerLabels(
-            user_id=scheduler_data.user_id,
-            project_id=scheduler_data.project_id,
-            node_id=scheduler_data.node_uuid,
-            product_name=scheduler_data.product_name,
-            simcore_user_agent=scheduler_data.request_simcore_user_agent,
-            swarm_stack_name=dynamic_services_scheduler_settings.SWARM_STACK_NAME,
-            memory_limit=ByteSize(0),  # this should get overwritten
-            cpu_limit=0,  # this should get overwritten
-        ).to_simcore_runtime_docker_labels()
-    )
+    standard_simcore_docker_labels: dict[DockerLabelKey, str] = SimcoreContainerLabels(
+        user_id=scheduler_data.user_id,
+        project_id=scheduler_data.project_id,
+        node_id=scheduler_data.node_uuid,
+        product_name=scheduler_data.product_name,
+        simcore_user_agent=scheduler_data.request_simcore_user_agent,
+        swarm_stack_name=dynamic_services_scheduler_settings.SWARM_STACK_NAME,
+        memory_limit=ByteSize(0),  # this should get overwritten
+        cpu_limit=0,  # this should get overwritten
+    ).to_simcore_runtime_docker_labels()
 
     service_labels: dict[str, str] = (
         {
