+++ services/director-v2/tests/unit/with_dbs/test_utils_dask.py
@@ -33,10 +33,10 @@ from models_library.api_schemas_storage.storage_schemas import (
     FileUploadLinks,
     FileUploadSchema,
 )
-from models_library.docker import to_simcore_runtime_docker_label_key
 from models_library.projects import ProjectID
 from models_library.projects_nodes_io import NodeID, SimCoreFileLink, SimcoreS3FileID
 from models_library.services import ServiceRunID
+from models_library.services_metadata_runtime import to_simcore_runtime_docker_label_key
 from models_library.users import UserID
 from pydantic import ByteSize, TypeAdapter
 from pydantic.networks import AnyUrl
